chore(sync): relocate account-deletion watch/sync cleanup out of sync/
user.service.ts's account-deletion and Google-disconnect paths (deleteCompassDataForUser,
stopGoogleCalendarSync, pruneGoogleData, handleLogoutCleanup) called into
packages/backend/src/sync/'s googleWatchService/syncRecords to stop Google watch channels
and clear the legacy sync collection. These aren't dead legacy-engine code: they're the
only thing that erases per-user rows in the sync/watch Mongo collections, and most current
users still have rows there from before the Sync service cutover, so this had to survive
sync/'s eventual wholesale deletion, not be dropped.

Relocated the self-contained pieces (stopWatch/stopWatches/deleteWatchesByUser - all pure
Mongo + Google API calls with zero sync/-internal dependencies) to
common/services/gcal/google-watch-cleanup.util.ts, matching the pattern already used for
gcal.client.ts/google-revoked.util.ts. google-watch.service.ts now imports and re-exports
them so its own remaining internal callers (all deleted alongside sync/ in the final pass)
keep working unchanged. Inlined the three one-line syncRecords Mongo deletions directly
into user.service.ts rather than keeping a wrapper module for three one-liners.

user.service.ts no longer imports anything from packages/backend/src/sync/.
